refactor: make initial changes for snail lib (shell interp)
